# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ControlHub SaaS Empresarial API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies/{companyId}/users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users in company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /companies/{companyId}/users/{userId}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # EMPLOYEES
...
  /companies/{companyId}/suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create supplier
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /companies/{companyId}/suppliers/{supplierId}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: supplierId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: supplierId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete supplier
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: supplierId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ATTENDANCE
...
  /companies/{companyId}/reports/finance:
    get:
      operationId: getFinanceReport
      tags: [reports]
      summary: Finance report
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Finance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceReport"
  /companies/{companyId}/reports/attendance:
    get:
      operationId: getAttendanceReport
      tags: [reports]
      summary: Attendance report
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attendance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceReport"
  /companies/{companyId}/reports/hr:
    get:
      operationId: getHrReport
      tags: [reports]
      summary: HR report
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: HR report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HrReport"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [superadmin, admin, hr, finance, supervisor, operator, employee]
        companyId:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, email, name, role, companyId, isActive, createdAt]
    CreateUserBody:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, hr, finance, supervisor, operator, employee]
      required: [email, name, password, role]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, hr, finance, supervisor, operator, employee]
        isActive:
          type: boolean
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        tradeName:
          type: string
          nullable: true
        ruc:
          type: string
          nullable: true
        contactName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        invoiceCount:
          type: integer
        totalBilled:
          type: number
        createdAt:
          type: string
      required: [id, companyId, name, invoiceCount, totalBilled, createdAt]
    CreateSupplierBody:
      type: object
      properties:
        name:
          type: string
        tradeName:
          type: string
        ruc:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
      required: [name]
    UpdateSupplierBody:
      type: object
      properties:
        name:
          type: string
        tradeName:
          type: string
        ruc:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
...
    FinanceReport:
      type: object
      properties:
        period:
          type: string
        summary:
          type: object
          properties:
            totalInvoiced:
              type: number
            totalPaid:
              type: number
            totalPending:
              type: number
            totalOverdue:
              type: number
          required: [totalInvoiced, totalPaid, totalPending, totalOverdue]
        byStatus:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
              amount:
                type: number
            required: [status, count, amount]
        bySupplier:
          type: array
          items:
            type: object
            properties:
              supplierName:
                type: string
              invoiceCount:
                type: integer
              totalAmount:
                type: number
            required: [supplierName, invoiceCount, totalAmount]
        monthlyTrend:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
              amount:
                type: number
            required: [month, amount]
      required: [period, summary, byStatus, bySupplier, monthlyTrend]
    AttendanceReport:
      type: object
      properties:
        period:
          type: string
        summary:
          type: object
          properties:
            totalDays:
              type: integer
            avgAttendanceRate:
              type: number
            totalHoursWorked:
              type: number
            totalOvertime:
              type: number
          required: [totalDays, avgAttendanceRate, totalHoursWorked, totalOvertime]
        byEmployee:
          type: array
          items:
            type: object
            properties:
              employeeName:
                type: string
              area:
                type: string
              presentDays:
                type: integer
              absentDays:
                type: integer
              lateDays:
                type: integer
              hoursWorked:
                type: number
            required: [employeeName, area, presentDays, absentDays, lateDays, hoursWorked]
        byArea:
          type: array
          items:
            type: object
            properties:
              area:
                type: string
              attendanceRate:
                type: number
              employeeCount:
                type: integer
            required: [area, attendanceRate, employeeCount]
      required: [period, summary, byEmployee, byArea]
    HrReport:
      type: object
      properties:
        summary:
          type: object
          properties:
            totalEmployees:
              type: integer
            byStatus:
              type: array
              items:
                type: object
                properties:
                  status:
                    type: string
                  count:
                    type: integer
                required: [status, count]
            byArea:
              type: array
              items:
                type: object
                properties:
                  area:
                    type: string
                  count:
                    type: integer
                required: [area, count]
            byContractType:
              type: array
              items:
                type: object
                properties:
                  contractType:
                    type: string
                  count:
                    type: integer
                required: [contractType, count]
          required: [totalEmployees, byStatus, byArea, byContractType]
      required: [summary]